Move controllers used by plugin to coredns
Following packages in `pkg` are used only by coredns plugin.
 * endpointslice
 * gateway
 * service
 * serviceimport

Move these to `coredns` package as that is more logical package
for them. This will simplify any changes to these as will no longer
need changes to come in two different PRs or require updating
`coredns/go.mod`.
