+++ coredns/endpointslice/map_test.go
@@ -23,9 +23,9 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/submariner-io/lighthouse/coredns/endpointslice"
+	"github.com/submariner-io/lighthouse/coredns/serviceimport"
 	lhconstants "github.com/submariner-io/lighthouse/pkg/constants"
-	"github.com/submariner-io/lighthouse/pkg/endpointslice"
-	"github.com/submariner-io/lighthouse/pkg/serviceimport"
 	discovery "k8s.io/api/discovery/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
