+++ coredns/plugin/handler.go
@@ -26,7 +26,7 @@ import (
 	"github.com/coredns/coredns/plugin"
 	"github.com/coredns/coredns/request"
 	"github.com/miekg/dns"
-	"github.com/submariner-io/lighthouse/pkg/serviceimport"
+	"github.com/submariner-io/lighthouse/coredns/serviceimport"
 )
 
 const PluginName = "lighthouse"
