+++ coredns/plugin/setup_internal_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/miekg/dns"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/submariner-io/lighthouse/pkg/endpointslice"
-	"github.com/submariner-io/lighthouse/pkg/gateway"
-	"github.com/submariner-io/lighthouse/pkg/serviceimport"
+	"github.com/submariner-io/lighthouse/coredns/endpointslice"
+	"github.com/submariner-io/lighthouse/coredns/gateway"
+	"github.com/submariner-io/lighthouse/coredns/serviceimport"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
